change: height of taskCard to be responsive based on length of message
