+++ Nudge-TaskManager/client/src/components/FullCard.tsx
@@ -1,7 +1,9 @@
-const FullCard = () => {
+import { StatTask } from "../interfaces/interfaces"
+
+const FullCard = ({TaskStat}:{TaskStat:StatTask}) => {
   return (
-    <div className="flex text-white border border-red-600 w-full h-full">
-      FullCard
+    <div className="flex text-white border border-red-600 w-full h-full justify-start">
+      <h1 className="text-white"> {TaskStat.status} </h1>
     </div>
   )
 }
