+++ Nudge-TaskManager/client/src/pages/Dashboard.tsx
@@ -6,21 +6,9 @@ import {
   IconBell,
 } from "@tabler/icons-react";
 import DashboardPage from "./DashboardContent";
-import { useState, createContext, Dispatch, SetStateAction } from "react";
-
-// Define the correct type for the context
-type ThemeContextType = {
-  selectDash: boolean;
-  setSelectDash: Dispatch<SetStateAction<boolean>>; // Correct type for the setter function
-};
-
-// Create the context with the default value
-const defaultThemeContext: ThemeContextType = {
-  selectDash: false,
-  setSelectDash: () => {}, // Default function
-};
-
-export const ThemeContext = createContext<ThemeContextType>(defaultThemeContext);
+import { useState } from "react";
+import NudgeLogo from "../assets/Group 1.svg";
+import { StatTask, TaskContent } from "../interfaces/interfaces";
 
 const DashBoard = () => {
   const teams = ["Team 1", "Team 2", "Team 3"];
@@ -28,7 +16,7 @@ const DashBoard = () => {
   const [changeTeam, setChangeTeam] = useState(false);
 
   // Toggle dashboard visibility
-  const DashBoard = () => {
+  const toggleDashboard = () => {
     setSelectDash(!selectDash);
   };
 
@@ -37,13 +25,118 @@ const DashBoard = () => {
     setChangeTeam(!changeTeam);
   };
 
+  const incompleteTasks: TaskContent[] = [
+    {
+      status: "Incomplete",
+      priority: "Low",
+      title: "Kent Task",
+      content: "Bomboclaat a;,sd;asd;a;da;ds;a,sd;a,ds;as,dla,sdl,asld,aldas,dla,sdla,sdla as,dla,sdla,dslaslmdlamasldlasdmlasmdlmasdmasdlmadslmalsmalsdmaldsmdddddddddddddddddddddddasdlamsdlmalsamsld",
+      assigned: ["Bryan","Vi", "Sky"],
+      comments: 3,
+      due: new Date('2024-5-11'), // Example due date
+    },
+    {
+        status: "Incomplete",
+        priority: "High",
+        title: "UI/UX Design",
+        content: "Wireframe needed in Figma akndddddddddddddddddddddddddddddddddddddddddddddddddddsssssssssssssssddddddddddddddddddddddddddddddddddddasd;as,d;a,sd;,a;d;a,sd;a,;ds,a;sd,;a,s;d,sdkandsknasdknaskdnansdkansdknaskdnakdsnklasndla,s;d,a;,sd;asd;a;da;ds;a,sd;a,ds;as,dla,sdl,asld,aldas,dla,sdla,sdla as,dla,sdla,dslaslmdlamasldlasdmlasmdlmasdmasdlmadslmalsmalsdmaldsmdddddddddddddddddddddddasdlamsdlmalsamsld",
+        assigned: ["Michael", "George", "Sky"],
+        comments: 5,
+        due: new Date('2024-11-15'), // Example due date
+    },
+    {
+        status: "Incomplete",
+        priority: "Medium",
+        title: "Content Creation",
+        content: "Draft new articles for the blog",
+        assigned: ["Ava", "Mia"],
+        comments: 2,
+        due: new Date('2024-11-20'), // Example due date
+    },
+    {
+        status: "Incomplete",
+        priority: "Low",
+        title: "Meeting Preparation",
+        content: "Organize agenda for client meeting",
+        assigned: ["Henry"],
+        comments: 0,
+        due: new Date('2024-11-25'), // Example due date
+    },
+];
+
+const inProgressTasks: TaskContent[] = [
+    {
+        status: "In Progress",
+        priority: "High",
+        title: "Security Audit",
+        content: "Conduct vulnerability assessments",
+        assigned: ["James", "Lucas"],
+        comments: 4,
+        due: new Date('2024-11-18'), // Example due date
+    },
+    {
+        status: "In Progress",
+        priority: "Medium",
+        title: "API Integration",
+        content: "Connect the frontend with backend services",
+        assigned: ["Sophia", "Oliver"],
+        comments: 3,
+        due: new Date('2024-11-22'), // Example due date
+    },
+    {
+        status: "In Progress",
+        priority: "Low",
+        title: "Marketing Plan",
+        content: "Develop a Q4 marketing strategy",
+        assigned: ["William", "Ella"],
+        comments: 1,
+        due: new Date('2024-11-30'), // Example due date
+    },
+];
+
+const completeTasks: TaskContent[] = [
+    {
+        status: "Complete",
+        priority: "High",
+        title: "Bug Fixes",
+        content: "Resolve critical issues in production",
+        assigned: [],
+        comments: 6,
+        due: new Date('2024-10-30'), // Example due date
+    },
+    {
+        status: "Complete",
+        priority: "Medium",
+        title: "Testing",
+        content: "Create test cases for new features",
+        assigned: ["Isabella", "Jack"],
+        comments: 3,
+        due: new Date('2024-10-25'), // Example due date
+    },
+    {
+        status: "Complete",
+        priority: "Low",
+        title: "Database Optimization",
+        content: "Analyze and improve query performance",
+        assigned: ["Emma"],
+        comments: 1,
+        due: new Date('2024-10-20'), // Example due date
+    },
+];
+
+const status: StatTask[] = [
+  { status: "Incomplete", Task: incompleteTasks},
+  { status: "In Progress", Task: inProgressTasks},
+  { status: "Complete", Task: completeTasks},
+];
+
   return (
-    <ThemeContext.Provider value={{ selectDash, setSelectDash }}>
+    // <DashThemeContext.Provider value={{ selectDash, setSelectDash }}>
       <div className="flex w-screen h-screen border-blue-600 overflow-y-hidden overflow-x-hidden">
         <div className="flex flex-col w-[304px] h-full shadow-custom-shadow border-r p-[32px] border-r-[#4B5D6A] bg-[#1A2329]">
           {/* Sidebar Content */}
           <div className="flex border-blue-600 h-max w-max">
-            <img src="../assets/Group 1.svg" alt="" width={152} />
+            <img src={NudgeLogo} alt="" width={152} />
           </div>
           <div className="mt-[29px] flex flex-col h-screen justify-between">
             {/* Main Menu */}
@@ -67,9 +160,9 @@ const DashBoard = () => {
                   color="#667988"
                   leftSection={<IconLayoutDashboard size="1rem" />}
                   className="flex items-center justify-start font-light"
-                  onClick={DashBoard}
+                  onClick={toggleDashboard}
                 >
-                  DashBoard
+                  Dashboard
                 </Button>
               </div>
               <div className="flex flex-col">
@@ -117,10 +210,10 @@ const DashBoard = () => {
               </div>
             </div>
           </div>
-          {selectDash && <DashboardPage />}
+          {selectDash && <DashboardPage StatTask={status}/>}
         </div>
       </div>
-    </ThemeContext.Provider>
+  //  </DashThemeContext.Provider>
   );
 };
